Merged files into main branch like structure.
Adjusted the file structure files so that Yanice's codes can be viewed via the React app. Additionally, I updated the README.md file with the current folder structure (on my device, so maybe it'll look different on other devices). Lastly, I uploaded the .md file for the Design And Implementation Document.